group number

+++ src/student/view/YearSemester.java
@@ -44,7 +44,6 @@ public class YearSemester extends javax.swing.JPanel {
         jPanel1 = new javax.swing.JPanel();
         jLabel1 = new javax.swing.JLabel();
         txt_acd_yer_and_sem = new javax.swing.JTextField();
-        btn_acd_yer_and_sem = new javax.swing.JButton();
         errorMsg = new javax.swing.JLabel();
         id = new javax.swing.JLabel();
         jLabel2 = new javax.swing.JLabel();
@@ -53,18 +52,12 @@ public class YearSemester extends javax.swing.JPanel {
         btn_back = new javax.swing.JButton();
         btn_delete_yer_and_sem = new javax.swing.JButton();
         btn_update_yer_sem = new javax.swing.JButton();
+        btn_acd_yer_and_sem = new javax.swing.JButton();
 
         jPanel1.setBackground(new java.awt.Color(255, 255, 255));
 
         jLabel1.setFont(new java.awt.Font("Tahoma", 1, 18)); // NOI18N
         jLabel1.setText("Academic Year And Semester");
-
-        btn_acd_yer_and_sem.setText("Submit");
-        btn_acd_yer_and_sem.addActionListener(new java.awt.event.ActionListener() {
-            public void actionPerformed(java.awt.event.ActionEvent evt) {
-                btn_acd_yer_and_semActionPerformed(evt);
-            }
-        });
 
         errorMsg.setForeground(java.awt.Color.red);
         errorMsg.setText("Error Message");
@@ -80,22 +73,17 @@ public class YearSemester extends javax.swing.JPanel {
         jPanel1Layout.setHorizontalGroup(
             jPanel1Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
             .addGroup(javax.swing.GroupLayout.Alignment.TRAILING, jPanel1Layout.createSequentialGroup()
-                .addGroup(jPanel1Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.TRAILING)
+                .addGap(29, 29, 29)
+                .addGroup(jPanel1Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
                     .addGroup(jPanel1Layout.createSequentialGroup()
-                        .addContainerGap(javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE)
-                        .addComponent(btn_acd_yer_and_sem))
-                    .addGroup(javax.swing.GroupLayout.Alignment.LEADING, jPanel1Layout.createSequentialGroup()
-                        .addGap(29, 29, 29)
-                        .addGroup(jPanel1Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
-                            .addGroup(jPanel1Layout.createSequentialGroup()
-                                .addComponent(errorMsg, javax.swing.GroupLayout.PREFERRED_SIZE, 319, javax.swing.GroupLayout.PREFERRED_SIZE)
-                                .addGap(0, 0, Short.MAX_VALUE))
-                            .addGroup(jPanel1Layout.createSequentialGroup()
-                                .addComponent(jLabel1)
-                                .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE)
-                                .addComponent(jLabel2)
-                                .addGap(48, 48, 48)
-                                .addComponent(txt_acd_yer_and_sem, javax.swing.GroupLayout.PREFERRED_SIZE, 78, javax.swing.GroupLayout.PREFERRED_SIZE)))))
+                        .addComponent(errorMsg, javax.swing.GroupLayout.PREFERRED_SIZE, 319, javax.swing.GroupLayout.PREFERRED_SIZE)
+                        .addGap(0, 0, Short.MAX_VALUE))
+                    .addGroup(jPanel1Layout.createSequentialGroup()
+                        .addComponent(jLabel1)
+                        .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE)
+                        .addComponent(jLabel2)
+                        .addGap(48, 48, 48)
+                        .addComponent(txt_acd_yer_and_sem, javax.swing.GroupLayout.PREFERRED_SIZE, 78, javax.swing.GroupLayout.PREFERRED_SIZE)))
                 .addGap(18, 18, 18)
                 .addComponent(id)
                 .addGap(48, 48, 48))
@@ -111,9 +99,7 @@ public class YearSemester extends javax.swing.JPanel {
                     .addComponent(jLabel2, javax.swing.GroupLayout.PREFERRED_SIZE, 9, javax.swing.GroupLayout.PREFERRED_SIZE))
                 .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED)
                 .addComponent(errorMsg)
-                .addGap(7, 7, 7)
-                .addComponent(btn_acd_yer_and_sem)
-                .addContainerGap(18, Short.MAX_VALUE))
+                .addContainerGap(48, Short.MAX_VALUE))
         );
 
         tbl_year.setModel(new javax.swing.table.DefaultTableModel(
@@ -152,6 +138,13 @@ public class YearSemester extends javax.swing.JPanel {
             }
         });
 
+        btn_acd_yer_and_sem.setText("Submit");
+        btn_acd_yer_and_sem.addActionListener(new java.awt.event.ActionListener() {
+            public void actionPerformed(java.awt.event.ActionEvent evt) {
+                btn_acd_yer_and_semActionPerformed(evt);
+            }
+        });
+
         javax.swing.GroupLayout layout = new javax.swing.GroupLayout(this);
         this.setLayout(layout);
         layout.setHorizontalGroup(
@@ -160,14 +153,16 @@ public class YearSemester extends javax.swing.JPanel {
                 .addGap(293, 293, 293)
                 .addGroup(layout.createParallelGroup(javax.swing.GroupLayout.Alignment.TRAILING)
                     .addGroup(layout.createSequentialGroup()
+                        .addComponent(btn_acd_yer_and_sem)
+                        .addGap(40, 40, 40)
                         .addComponent(btn_update_yer_sem)
-                        .addGap(49, 49, 49)
+                        .addGap(38, 38, 38)
                         .addComponent(btn_delete_yer_and_sem)
-                        .addGap(48, 48, 48)
+                        .addGap(30, 30, 30)
                         .addComponent(btn_back))
                     .addGroup(layout.createParallelGroup(javax.swing.GroupLayout.Alignment.TRAILING, false)
                         .addComponent(jScrollPane1, javax.swing.GroupLayout.DEFAULT_SIZE, 477, Short.MAX_VALUE)
-                        .addComponent(jPanel1, javax.swing.GroupLayout.PREFERRED_SIZE, 0, Short.MAX_VALUE)))
+                        .addComponent(jPanel1, javax.swing.GroupLayout.PREFERRED_SIZE, 477, Short.MAX_VALUE)))
                 .addContainerGap(354, Short.MAX_VALUE))
         );
         layout.setVerticalGroup(
@@ -181,7 +176,8 @@ public class YearSemester extends javax.swing.JPanel {
                 .addGroup(layout.createParallelGroup(javax.swing.GroupLayout.Alignment.BASELINE)
                     .addComponent(btn_delete_yer_and_sem)
                     .addComponent(btn_back)
-                    .addComponent(btn_update_yer_sem))
+                    .addComponent(btn_update_yer_sem)
+                    .addComponent(btn_acd_yer_and_sem))
                 .addContainerGap(213, Short.MAX_VALUE))
         );
     }// </editor-fold>//GEN-END:initComponents
@@ -192,13 +188,15 @@ public class YearSemester extends javax.swing.JPanel {
 
             if (inputDataValidator() == true && allReady() == true) {
 
-                String yer_sem = txt_acd_yer_and_sem.getText().trim().toUpperCase();
+                errorMsg.setVisible(false);
 
+                String yer_sem = txt_acd_yer_and_sem.getText().trim().toUpperCase();
+                showYnSList();
                 int i = yc.addHandleClick(yer_sem);
                 if (i != 0) {
                     JOptionPane.showMessageDialog(jPanel1, "Successfully Addeda!", "Done", JOptionPane.PLAIN_MESSAGE);
                     txt_acd_yer_and_sem.setText("");
-                    showYnSList();
+
                 } else {
                     JOptionPane.showMessageDialog(jPanel1, "Failed!", "Error", JOptionPane.ERROR_MESSAGE);
                 }
@@ -241,24 +239,33 @@ public class YearSemester extends javax.swing.JPanel {
         String idYS = id.getText();
         int final_id = Integer.parseInt(idYS);
 
-        JOptionPane.showMessageDialog(jPanel1, "Are You Sure Delete", "Delete", JOptionPane.OK_OPTION);
-        
-        try {
+        int v = JOptionPane.showConfirmDialog(jPanel1, "Are You Sure Delete", "Delete", JOptionPane.YES_NO_OPTION);
 
-            int i = yc.deleteHandleClick(final_id);
+        if (v == JOptionPane.YES_OPTION) {
 
-            if (i != 0) {
+            try {
+
+                int i = yc.deleteHandleClick(final_id);
+
+                if (i != 0) {
 //                    JOptionPane.showMessageDialog(jPanel1, "Successfully Updated!", "Done", JOptionPane.PLAIN_MESSAGE);
-                JOptionPane.showMessageDialog(jPanel1, "Successfully Deleted!");
-                txt_acd_yer_and_sem.setText("");
-                showYnSList();
-            } else {
-                JOptionPane.showMessageDialog(jPanel1, "Failed!", "Error", JOptionPane.ERROR_MESSAGE);
+                    JOptionPane.showMessageDialog(jPanel1, "Successfully Deleted!");
+                    txt_acd_yer_and_sem.setText("");
+                    showYnSList();
+
+                } else {
+                    JOptionPane.showMessageDialog(jPanel1, "Failed!", "Error", JOptionPane.ERROR_MESSAGE);
+                }
+
+            } catch (SQLException ex) {
+
+                Logger.getLogger(YearSemester.class.getName()).log(Level.SEVERE, null, ex);
+
             }
 
-        } catch (SQLException ex) {
+        } else if (v == JOptionPane.NO_OPTION) {
 
-            Logger.getLogger(YearSemester.class.getName()).log(Level.SEVERE, null, ex);
+        } else {
 
         }
 
@@ -272,6 +279,8 @@ public class YearSemester extends javax.swing.JPanel {
         try {
 
             if (inputDataValidator() == true && allReady() == true) {
+
+                errorMsg.setVisible(false);
 
                 String yer_sem = txt_acd_yer_and_sem.getText().trim().toUpperCase();
